==> notebook/00-default-hyperparams.ipynb <==
import numpy as np

hyperparams = np.array([
["fs", "combine_fs",{
        "combine_fs__percentile": [5, 10, 20, 30, 40, 50]}],
["fs", "rfe_rf_fs",{
        "rfe_rf_fs__step": [0.1],
        "rfe_rf_fs__n_features_to_select": [5,10,15,20]}],
["fs", "lasso_fs",{
        "lasso_fs__estimator__C": [0.001, 0.01, 0.1, 1]}],
["cls", "knn",{
        "knn__n_neighbors": [3, 5, 9, 11],
        "knn__weights": ["uniform", "distance"]}],
["cls", "logReg",{
        "logReg__C": [0.00001, 0.0001, 0.001, 0.01, 0.1, 1, 3,10],
        "logReg__class_weight": ["balanced"],
        "logReg__penalty": ["l1", "l2"]}],
["cls", "svmRBF",{
        "svmRBF__C": [0.01, 0.1, 0.5, 1, 5, 10, 30, 50, 100],
        "svmRBF__gamma": [0.0001, 0.001, 0.01, 0.1, 1, 5],
        'svmRBF__class_weight': ['balanced']}], 
["cls", "rf",{
        "rf__n_estimators": [50,100,200,300,500],
        "rf__criterion": ["entropy", "gini"],
        'rf__max_depth' : [None,2,4,6,8],
        'rf__class_weight':["balanced_subsample"]}],
["cls", "nn",{
        'nn__alpha': [1e-5,0.00001,0.0001,0.001,0.01,0.1,1,3,5,10],
        'nn__hidden_layer_sizes':[(30,),(50,),(70,),(100,),(150,),
                                  (30,30),(50,50),(70,70),(100,100),
                                  (30,30,30),(50,50,50),(70,70,70)
                                 ]}],
["cls", "gbt",{
        'gbt__n_estimators': [300,400,500],
        'gbt__learning_rate': [0.1,0.01,0.001],
        'gbt__max_depth' : [None,8,10,12]}],
["after", "sm_smote",{"sm_smote__k_neighbors":[3,4,5]}]

])

np.save("../src/default_hyperparams.npy",hyperparams)

print hyperparams.shape
print hyperparams[0,1]
