==> notebook/05-results.ipynb <==
%reset -f

from IPython.display import display, HTML
from sklearn import metrics
from sklearn import cross_validation
from scipy import interp
import pandas as pd
import numpy as np
import os

import sys
sys.path.insert(1,"../../src/")
from TypeFeatImputer import TypeFeatImputer
from i_score_parallel import i_score
from TypeFeatFS import DiscreteFS, ContinuousFS
from TypeFeatFilter import DiscreteFilter, ContinuousFilter
from OutlierFiltering import OutlierFiltering
from typeFeat_score import typeFeat_score

from pylab import *
import seaborn as sns
import matplotlib.pyplot as plt


np.set_printoptions(suppress=True)
pd.options.display.float_format = '{:,.4f}'.format
plt.style.use('classic')

%matplotlib inline

ddf = []
pathDir = os.path.join('resources','results') 
path, dirs, files = os.walk(pathDir).next()
file_count = len(files)

for f in files:
    if "summary" not in str(f):
        ddf.append(pd.read_pickle(os.path.join('resources','results', str(f) )))
    
dfAll = pd.concat(ddf)
print dfAll.shape

dfAll.cv_f1_mean = pd.to_numeric(dfAll.cv_f1_mean)
dfAll.cv_auc_mean = pd.to_numeric(dfAll.cv_auc_mean)
dfAll.cv_sens_mean = pd.to_numeric(dfAll.cv_sens_mean)
dfAll.cv_spec_mean = pd.to_numeric(dfAll.cv_spec_mean)

dfAll.cv_f1_std = pd.to_numeric(dfAll.cv_f1_std)
dfAll.cv_auc_std = pd.to_numeric(dfAll.cv_auc_std)
dfAll.cv_sens_std = pd.to_numeric(dfAll.cv_sens_std)
dfAll.cv_spec_std = pd.to_numeric(dfAll.cv_spec_std)
dfAll.exp = pd.to_numeric(dfAll.exp)

dfAll["cv_auc"] = dfAll["cv_auc_mean"].round(3).astype(str).str.cat(dfAll["cv_auc_std"].round(3).astype(str), sep="+/-")
dfAll["cv_f1"] = dfAll["cv_f1_mean"].round(3).astype(str).str.cat(dfAll["cv_f1_std"].round(3).astype(str), sep="+/-")
dfAll["cv_sens"] = dfAll["cv_sens_mean"].round(3).astype(str).str.cat(dfAll["cv_sens_std"].round(3).astype(str), sep="+/-")
dfAll["cv_spec"] = dfAll["cv_spec_mean"].round(3).astype(str).str.cat(dfAll["cv_spec_std"].round(3).astype(str), sep="+/-")
dfAll["disease"] = dfAll["name"].str.slice(0,dfAll["name"].str.find("_").values[0])

print
print dfAll.columns.tolist()
print dfAll.name.unique()
print dfAll["disease"].unique()
print dfAll.shape

typeEncounter = "last" # ['first','last']
typeHypothesis = "all_readmisssion_vs_none" # ['all_readmisssion_vs_none','early_readmission_vs_none']
typeDataFeatures = "extended_extra_diag_1" # ["reduced","extended','extended_extra','extended_extra_diag_1','extended_extra_diag_3']
    #Extended_Extra_diag_1 -> Extended extra columns with disease in diag_3
    #Extended_Extra_diag_1 -> Extended extra columns with disease in diag_1
    #Extended_Extra -> Extended extra of columns
    #Extended -> Extended columns
    #Reduced -> minimum set of columns 
typeDataExperiment = "all" #["all", "disease"] 
    #all -> No filter disease column
    #disease -> Filter/Remove disease column and keep only rows with diagnosis column == 1  

name = typeDataExperiment + "_" + typeDataFeatures + "_" +  typeEncounter + "_" + typeHypothesis

print name

#Sort and list all pipelines
dfAll.sort_values(["cv_auc_mean"], ascending=False,inplace=True)
condition = np.logical_and(dfAll['name'].str.contains(typeHypothesis), 
                           dfAll.disease != typeDataExperiment)
print dfAll[condition].shape
print np.sum(dfAll[condition]["time"])


dfAll[condition][["disease","name","fs","metric","cls","params",
    "tr_auc","tr_sens","tr_spec",
    "cv_auc_mean","cv_f1_mean","cv_sens_mean","cv_spec_mean",
    "test_auc","test_f1","test_sens","test_spec"]]

#Filter pipelines less performance than random
dfAllAux = dfAll[condition]
dfAllAux[[
    "name","fs","metric","params",
    "tr_auc","tr_sens","tr_spec",
    "cv_auc_mean","cv_f1_mean","cv_sens_mean","cv_spec_mean",
    "test_auc","test_f1","test_sens","test_spec"]]

dfAllAux[[
    "metric","fs","cls","params",
    "tr_auc","tr_sens","tr_spec",
    "cv_auc_mean","cv_f1_mean","cv_sens_mean","cv_spec_mean",
    "test_auc","test_f1","test_sens","test_spec"]].to_latex()

#### Grouped by Diseases

dfAllAux = dfAllAux.sort_values(["disease","cv_auc_mean"], ascending=False,inplace=False)
dfAllAux = dfAllAux[np.logical_and(dfAllAux["cv_sens_mean"] > 0.5,dfAllAux['cv_spec_mean']> 0.5)]
dfAllAux = dfAllAux.groupby(["disease"]).first()
dfAllAux[["sm","fs","metric","cls","params",
    "tr_auc","tr_sens","tr_spec",
    "cv_auc_mean","cv_f1_mean","cv_sens_mean","cv_spec_mean",
    "test_auc","test_f1","test_sens","test_spec"]]

print dfAllAux[["metric","cls","params","tr_auc","tr_sens","tr_spec",
    "cv_auc_mean","cv_f1_mean","cv_sens_mean","cv_spec_mean","test_auc",
                "test_f1","test_sens","test_spec"]].to_latex()

#### Save results

import time

fname = "summary" + "_" + name + "_" + time.strftime("%Y%m%d-%H%M%S") + ".pkl"
dfAllAux.to_pickle(os.path.join(pathDir, fname))
print fname

dd = pd.read_pickle(os.path.join(pathDir,fname))
print dd.columns
